requirements.txt
Django
python-decouple
psycopg2
Pillow
djangorestframework
requests
drf-spectacular
django-silk
redis[hiredis]
django-redis
djangorestframework-simplejwt[crypto]
pika
python-dotenv
pymongo
nltk